rename second public circuit to app_logic
