+++ yarn-project/noir-protocol-circuits/src/Nargo.toml
@@ -11,8 +11,8 @@ members = [
     "crates/public-kernel-lib",
     "crates/public-kernel-setup",
     "crates/public-kernel-setup-simulated",
-    "crates/public-kernel-public-previous",
-    "crates/public-kernel-public-previous-simulated",
+    "crates/public-kernel-app-logic",
+    "crates/public-kernel-app-logic-simulated",
     "crates/rollup-lib",
     "crates/rollup-merge",
     "crates/rollup-base",
