+++ yarn-project/noir-protocol-circuits/src/crates/public-kernel-app-logic-simulated/Nargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "public_kernel_public_previous"
+name = "public_kernel_app_logic_simulated"
 type = "bin"
 authors = [""]
 compiler_version = ">=0.18.0"
