+++ yarn-project/noir-protocol-circuits/src/crates/public-kernel-app-logic/Nargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "public_kernel_public_previous_simulated"
+name = "public_kernel_app_logic"
 type = "bin"
 authors = [""]
 compiler_version = ">=0.18.0"
