+++ yarn-project/noir-protocol-circuits/src/crates/public-kernel-app-logic/src/main.nr
@@ -1,0 +1,6 @@
+use dep::public_kernel_lib::{PublicKernelAppLogicCircuitPrivateInputs};
+use dep::types::{PublicKernelCircuitPublicInputs};
+
+fn main(input: PublicKernelAppLogicCircuitPrivateInputs) -> distinct pub PublicKernelCircuitPublicInputs {
+    input.public_kernel_app_logic()
+}
