+++ yarn-project/noir-protocol-circuits/src/crates/public-kernel-lib/src/lib.nr
@@ -5,7 +5,7 @@ mod hash;
 mod utils;
 
 mod public_kernel_setup;
-mod public_kernel_public_previous;
+mod public_kernel_app_logic;
 
 use public_kernel_setup::PublicKernelSetupCircuitPrivateInputs;
-use public_kernel_public_previous::PublicKernelPublicPreviousInputs;
+use public_kernel_app_logic::PublicKernelAppLogicCircuitPrivateInputs;
