--- yarn-project/noir-protocol-circuits/src/crates/public-kernel-public-previous/src/main.nr
@@ -1,6 +1,0 @@
-use dep::public_kernel_lib::{PublicKernelPublicPreviousInputs};
-use dep::types::{PublicKernelCircuitPublicInputs};
-
-fn main(input: PublicKernelPublicPreviousInputs) -> distinct pub PublicKernelCircuitPublicInputs {
-    input.public_kernel_public_previous()
-}
